Update 98_FAQ.md
Added the following sections to the top of the FAQ based on this webinar and set of slides:
https://www.youtube.com/watch?v=dDuP-Znso5Y
https://docs.google.com/presentation/d/1H4C_cBU9BE9EYFNyK8YBLQObzvq9iDztLWrqRIQTmgc/edit#slide=id.p

## What is ReproSchema
## What challenges is ReproSchema trying to solve?
## Why should I use Reproschema?
## What are the key elements of ReproSchema?

These primarily reflect my growing understanding of this system. I am actually not 100% sure yet who the audience is for this FAQ so I am not sure if this initial draft of edits would be useful or not.